+++ sdk/core/azure-core/src/main/java/com/azure/core/util/Configuration.java
@@ -113,6 +113,11 @@ public class Configuration implements Cloneable {
     public static final String PROPERTY_AZURE_CLIENT_CERTIFICATE_PASSWORD = "AZURE_CLIENT_CERTIFICATE_PASSWORD";
 
     /**
+     * Flag to enable sending the certificate chain in x5c header to support subject name / issuer based authentication.
+     */
+    public static final String PROPERTY_AZURE_CLIENT_SEND_CERTIFICATE_CHAIN = "AZURE_CLIENT_SEND_CERTIFICATE_CHAIN";
+
+    /**
      * Flag to disable the CP1 client capabilities in Azure Identity Token credentials.
      */
     public static final String PROPERTY_AZURE_IDENTITY_DISABLE_CP1 = "AZURE_IDENTITY_DISABLE_CP1";